pyproject.toml
[build-system]
requires = [
    "setuptools==82.0.0",
    "setuptools_scm==9.2.2",
    "setuptools_dynamic_dependencies==1.0.0",
]
build-backend = "setuptools.build_meta"

[project]
version = "0.0.1"
name = "toga-bitmap-view"
description = "A Python native, OS native GUI toolkit."
readme = "README.md"
requires-python = ">= 3.10"
license = "BSD-3-Clause"
license-files = [
    "LICENSE"
]
authors = [
    {name="Russell Keith-Magee", email="russell@keith-magee.com"},
    {name="Corran Webster", email="cwebster@unital.dev"},
]
maintainers = [
    {name="Corran Webster", email="cwebster@unital.dev"},
]
keywords = [
    "gui",
    "widget",
    "cross-platform",
    "toga",
    "desktop",
    "mobile",
    "web",
    "macOS",
    "cocoa",
    "iOS",
    "android",
    "windows",
    "winforms",
    "linux",
    "freeBSD",
    "gtk",
    "console",
    "web",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Software Development :: Widget Sets",
]
dependencies=[
    "toga_core",
]

# New Widgets
[project.entry-points."togax_bitmap_view.backend.toga_cocoa"]
BitmapView = "toga_bitmap_view.cocoa_bitmap_view:BitmapView"

[project.entry-points."togax_bitmap_view.backend.toga_gtk"]
BitmapView = "toga_bitmap_view.gtk_bitmap_view:BitmapView"

[dependency-groups]
# Extras used by developers *of* Toga are pinned to specific versions to
# ensure environment consistency.
test = [
    "coverage[toml] == 7.13.4",
    "coverage-conditional-plugin == 0.9.0",
    "Pillow == 12.1.1",
    "pytest == 9.0.2",
    "pytest-asyncio == 1.3.0",
    "pytest-freezer == 0.4.9",
    "pytest-xdist == 3.8.0",
    "setuptools-scm==9.2.2",
]

[project.urls]
#Documentation = "https://toga.beeware.org/"
Tracker = "https://github.com/corranwebster/toga_bitmap_view/issues"
Source = "https://github.com/corranwebster/toga_bitmap_view"
#Changelog = "https://toga.beeware.org/en/stable/background/project/releases"

[tool.setuptools_scm]
root = ".."

[tool.coverage.run]
parallel = true
branch = true
relative_files = true

# See notes in the root pyproject.toml file.
source = ["src"]
source_pkgs = ["toga_bitmap_view"]

[tool.coverage.paths]
source = [
    "src/toga_bitmap_view",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "error",
]
